:bug: Fix bugs for compatibility

+++ environments/environment_abstract.py
@@ -17,7 +17,8 @@ class State(ABC):
 
 class Environment(ABC):
     def __init__(self):
-        self.dtype = np.float
+        # self.dtype = np.float
+        self.dtype = np.float32
         self.fixed_actions: bool = True
 
     @abstractmethod
